Optimize GRA_bitmap drawing by drawing it all at once
wxDC::DrawPoint() in a loop is the slowest possible way to draw bitmaps,
so switch to a much faster way using wxDC::DrawBitmap().

The bitmap is currently initialized from wxImage, to keep the code as
simple as possible, while already being fast enough, but if we wanted to
optimize it further, we could use raw bitmap access via pixel iterator
classes.

We could also optimize bitmap_ to store the data in the same format as
wxImage does (or even use wxImage for RGB values storage), as then we
wouldn't have to do conversion to wxImage at all here.