Stop using recursive automake
There is no need to recurse into the subdirectories and this just
complicates things needlessly, as well as making (especially highly
parallel) builds somewhat slower than optimal, so move all the targets
from the sub-makefiles to the main src/Makefile.am. Note that this
requires enabling automake subdir-objects option, but there is no real
problem with doing this.

In fact, there is no need to even have separate libraries for each
subdirectory contents, but keep them for now, in case we do need them in
the future and just to avoid changing too many things at once. But just
compiling all the files together directly into the main application
would be even simpler and (slightly) faster.