Restore RefreshGraphics() call in CMD_density::Execute()
Removing it in 9225c00 (Avoid an apparently unnecessary refresh in
CMD_density, 2021-04-22) broke background redrawing, as this is only
done during full repaint in GRA_window::Draw(), so restore this call
and, instead, to avoid drawing all the density plots twice, don't draw
them on screen at all in Execute() but only draw them on an off-screen
wxMemoryDC just to check that we get no errors while drawing.

Notice that this is still quite inefficient as we go through everything
(notably conversion to wxImage and wxBitmap) but actual drawing, and it
would be best to avoid doing this entirely, if possible.