Remove multitude of useless try/catch clauses
Catching an exception just to rethrow it is completely useless, not
catching it at all results in exactly the same behaviour, so don't do
it and just remove all of them.

No real changes, but the code is much shorter and clear now.

This commit is best viewed ignoring whitespace-only changes.