Save and restore window geometry automatically
Use "persistent windows" support in wx to do it in a single line of
code, rather than doing it manually using wxConfig.

In addition to being much simpler to do, this has the advantage of
taking into account many cases not previously handled at all (e.g. we
now restore iconized and maximized windows correctly) as well as working
around various problems due to having to deal with GTK and WM
interaction which is now done inside wxGTK itself.

Note that wxConfig is still used to save/restore Extrema-specific
options. We could plug into wxPersistentManager to do it too, but for
now it's not clear if it's worth it, so don't do it. It probably should
be done for restoring the active notebook page, as this is something
that is already supported by wx itself.

Also note that the config file format has changed, but it doesn't seem
worth it to import the old settings into the new ones, so keep things
simple by not doing it -- having to reposition the windows once, when
upgrading to a new version, doesn't seem like a huge problem.