Fix a multitude of modal dialog leaks
Don't allocate modal dialogs on the heap without freeing them, they're
supposed to be deleted by the caller because they are almost invariably
created on the stack and so can't be destroyed by wxWidgets itself.

Just do this for the dialogs that we really need to create and replace a
bunch of wxMessageDialogs with a simple call to a wrapper wxMessageBox()
function which creates and shows the dialog modally inside it.

There should be no changes in behaviour, but no more dialog leaks.

See #9.