Save all scripts when the execute dialog is destroyed, not closed
The dialog is always destroyed (unless the program dies before), but
it's not closed if it's destroyed implicitly, due to the program exit,
rather than the user explicitly closing it, and we want to always save
the scripts entered in it, so do it in the dtor rather than EVT_CLOSE
handler.

Fixes #11.