Add ExGlobals::GetGraphicsPage() helper and use instead of casts
This returns the same window as the existing GetwxWindow(), but of the
right type, i.e. as a GraphicsPage pointer.

This allows to get rid of some casts in the code using GraphicsPage
(where static_cast<> or dynamic_cast<> were used in apparently random
way, so this also gets rid of the question about which one of them
should be used) and will be useful in the upcoming commits, but no real
changes so far.