Remove unused wxClientDC objects
They became unnecessary after commenting out Draw() calls using them in
5aab3e2 (propose removal of all ->Draw() calls, 2021-04-30), so simply
remove these objects themselves too.

Also remove all calls to Draw() from commands code, including the
already commented out ones and other similar ones that were not affected
by the changes of the commit above.

Also remove the ones using dummy DC instead of client DC since the
changes of ba9a960 (Restore RefreshGraphics() call in
CMD_density::Execute(), 2021-04-22), as they all are definitely
unnecessary from the point of view of screen updating (as this will be
(re)done when RefreshGraphics() is called anyhow) and if they are really
needed for error-checking, we should do it in AddDrawableObject() and
for all kinds of objects and not just this one, as otherwise we'd have
to reproduce this hack with a dummy memory DC in many places, which
definitely doesn't seem desirable.

Note that more, probably also unnecessary, calls to Draw() remain in
graphics code.