Replace all uses of wxClientDC with wxMemoryDC
Draw on off-screen bitmap which will be drawn on the screen later,
rather than on the screen directly because this works even with GTK 3
(and even when using Wayland) and Mac, where wxClientDC can't be used
for drawing at all, and because this avoids visible changes even on the
platforms where it still does work.

Do this by adding a new GraphicsDC class which uses the backing store
bitmap inside GraphicsPage and refreshes it automatically when it is
used. This allows to just replace the existing occurrences of wxClientDC
with GraphicsDC without making any further changes.

This is still much less efficient than it could be because for now the
changes are minimal and try to exactly preserve the current behaviour,
and should be further optimized by avoiding unnecessary redraws and
refreshes later.