Redraw if necessary before generating the PostScript file
All graphics objects must be drawn on an actual wxDC before being
rendered as PostScript because some of their metrics are only updated
when their Draw() method is called -- notably, for the axes, the width
is only computed then, and not doing it resulted in wrong PostScript
output if there was no refresh before it was generated, as happened if
it was done by hardcopy\postscript command in a script.

Fix this by simply calling GetBackingStore() to ensure that everything
is drawn, if necessary.

Closes #25.