Fix compilation problems in MinGW build due to using toascii()
When using this compiler CRT, toascii() can't be used with wxUniChar, as
it's just a macro expanding into "x & 0x7f", where "x" is its argument,
and this expression is ambiguous for "x" of type wxUniChar.

This could be solved in several different ways, but using toascii() is,
generally speaking, not a great idea, so add a new TryConvertToAscii()
function which makes it more clear that the conversion may fail and use
it only after confirming that returning 0 for non-ASCII characters is
actually correctly in the places where it is used now.

In the new code, explicitly using wxUniChar::IsAscii() or, even better,
not special-casing behaviour for ASCII, would be preferable.