Fix listbox styles used in "Write XXX" forms
wxLB_EXTENDED can't be used with wxLB_MULTIPLE, these styles are
exclusive and using both of them triggers a debugging assert in wx.
We probably want just wxLB_MULTIPLE here, so leave this style only.

wxLB_NEEDED_SB can still be useful but is perfectly useless (it's
defined as 0 anyhow), so remove it.