Use intermediate panels in all frames
This fixes problems with frame background showing through the place not
occupied by any controls under MSW and also fixes TAB navigation, which
doesn't work outside of a panel.

In most places, just creating a local wxPanel variable and using it as
the parent was enough. However in a couple of them, "this" of
wxFrame-derived type was used both as the window parent and as an
Extrema-specific variable and in these cases GetMainPanel() had to be
added, so that it could be used as the parent at wx level while still
using "this" for the other use.