Make script execution dialog work under MSW
It didn't work previously as backslashes in the full script file paths
were misinterpreted by ParseLine::ParseIt().

Avoid the problem by using just the file name in the command line, this
is enough because the code already uses ExGlobals::SetCurrentPath() to
set the path.

This is not a complete fix as any full paths coming from elsewhere
(including interactive entry in the main window) still don't work under
MSW.