Replace filled_ring.pcm with filledring.pcm
The two files were almost identical and never modified since 30bcb71
(initial import, 2007-06-14), so it doesn't seem to be worth having two
of them.

Using underscores is a more usual convention, so keep the file using it,
but use the contents of the old filledring.pcm in it, as it seems to be
a more evolved version of the same file.