Fix wrong initial size of the visualization window graphics
We need to set the aspect ration before laying out the pages as
ResetPages() function used by layout code uses the current aspect
ration.

We also don't need to call SetMonitorLimits() at all, as this is already
done by ResetPages() and, in fact, calling it here replaced the correct
limits set there with wrong ones (as the size here is the frame size and
not the graphics page size at all).