Use the best available approximation for Extrema fonts
Check if the font we want to use is available, try using "Free"-prefixed
version (which corresponds to the names of the fonts distributed with
Extrema) otherwise and, finally, if it's still not available, use the
default font of the corresponding font family for the current platform.

This should result in better font fidelity under Windows, both when
FreeXXX fonts are available and when they are not.